Rename API package and move SC v1 files
