+++ api/src/main/kotlin/com/meamoria/lexurgy/api/sc/v1/Scv1Request.kt
@@ -1,4 +1,4 @@
-package com.meamoria.lexurgy.services
+package com.meamoria.lexurgy.api.sc.v1
 
 import kotlinx.serialization.Serializable
