+++ api/src/main/kotlin/com/meamoria/lexurgy/api/sc/v1/Scv1Response.kt
@@ -1,4 +1,4 @@
-package com.meamoria.lexurgy.services
+package com.meamoria.lexurgy.api.sc.v1
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
