+++ api/src/test/kotlin/com/meamoria/lexurgy/api/TestApi.kt
@@ -1,4 +1,4 @@
-package com.meamoria.lexurgy.services
+package com.meamoria.lexurgy.api
 
 import io.ktor.client.*
 import io.ktor.client.request.*
