+++ api/src/test/kotlin/com/meamoria/lexurgy/api/VersionTest.kt
@@ -1,4 +1,4 @@
-package com.meamoria.lexurgy.services
+package com.meamoria.lexurgy.api
 
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
